chatbot/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.27.0
pydantic==2.6.2
python-multipart==0.0.9